feat(logs): highlight filtering keywords in logger and message
